Removed flawed ht_del from the hashtable implementation.
The function as it is written is flawed. It is not enough to empty the
cell when removing an entry from the hashtable (using linear probing).

See:
https://en.wikipedia.org/wiki/Linear_probing#Deletion

Since, there is no need to delete any entry in the hashtable for this
project. (i.e., ht_del is never actually called), we remove this
function, for now, to prevent any future use of a flawed function.

Again thanks to @emanuele6 for pointing out this issue.